I tried
Directions:
Add code to attendance.py that allows the teacher to enter the letter L (or l) to begin marking students late. When this happens, the current time should be added to both the student's 'attendance' AND 'lates' lists.
Update:
The code will mark student late when teacher enters l or L and the time will be recorded in the attendance shelve but I'm not sure if a new shelve would need to be built so that it will mark the student's late days. So I just tried to add it to the shelve that was already built in. I also don't know if I should use if or elif..